Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -6,7 +6,7 @@ int main(){
     scanf("%d",&b);
     scanf("%c",&c);
     int x = a*b;
-    float y = a/b;
+    int y = a/b;
     int z = a+ b;
     int r = a-b;
     switch(c){
@@ -14,7 +14,7 @@ int main(){
             printf("%i",x);
             break;
         case '/':
-            printf("%.2f",y);
+            printf("%d",y);
             break;
         case '+':
             printf("%i",z);
